+++ src/main/java/fll/web/report/TopScoreReportPerAwardGroupPdf.java
@@ -14,9 +14,11 @@ import java.util.Map;
 
 import javax.sql.DataSource;
 
+import fll.ScoreStandardization;
 import fll.Tournament;
 import fll.util.FLLRuntimeException;
 import fll.web.ApplicationAttributes;
+import fll.web.TournamentData;
 import fll.web.scoreboard.Top10;
 import fll.xml.ChallengeDescription;
 import jakarta.servlet.ServletContext;
@@ -38,24 +40,22 @@ public class TopScoreReportPerAwardGroupPdf extends TopScoreReportPdf {
                                 final ServletContext application,
                                 final HttpSession session)
       throws IOException, ServletException {
-    if (PromptSummarizeScores.checkIfSummaryUpdated(request, response, application, session,
-                                                    "/report/TopScoreReportPerAwardGroupPdf")) {
-      return;
-    }
-
     response.reset();
     response.setContentType("application/pdf");
     response.setHeader("Content-Disposition", "filename=TopScoreReportPerAwardGroup.pdf");
 
-    final DataSource datasource = ApplicationAttributes.getDataSource(application);
+    final TournamentData tournamentData = ApplicationAttributes.getTournamentData(application);
+    final DataSource datasource = tournamentData.getDataSource();
     final ChallengeDescription description = ApplicationAttributes.getChallengeDescription(application);
     try (Connection connection = datasource.getConnection()) {
-      final Tournament tournament = Tournament.getCurrentTournament(connection);
-      final ChallengeDescription challengeDescription = ApplicationAttributes.getChallengeDescription(application);
+      ScoreStandardization.computeSummarizedScoresIfNeeded(connection, description,
+                                                          tournamentData.getCurrentTournament());
+
+      final Tournament tournament = tournamentData.getCurrentTournament();
 
       final Map<String, List<Top10.ScoreEntry>> scores = Top10.getTableAsMapByAwardGroup(connection, description, true,
                                                                                          false);
-      outputReport(response.getOutputStream(), challengeDescription, tournament, "Award Group", scores);
+      outputReport(response.getOutputStream(), description, tournament, "Award Group", scores);
     } catch (final SQLException e) {
       throw new FLLRuntimeException("Error talking to the database", e);
     }